fix(core): failing tests
