+++ packages/nx/src/command-line/release/version/release-group-processor.spec.ts
@@ -1675,7 +1675,7 @@ describe('ReleaseGroupProcessor', () => {
   });
 
   describe('versionData', () => {
-    it('should populate versionData even when projects are not versioned', async () => {
+    it('should populate versionData even ~when projects are not versioned', async () => {
       const {
         nxReleaseConfig,
         projectGraph,
@@ -1723,11 +1723,13 @@ describe('ReleaseGroupProcessor', () => {
           "libtest1": {
             "currentVersion": "4.5.0",
             "dependentProjects": [],
+            "dockerVersion": null,
             "newVersion": null,
           },
           "my-nest-app": {
             "currentVersion": "1.0.0",
             "dependentProjects": [],
+            "dockerVersion": null,
             "newVersion": null,
           },
         }
